created GetHeadPose function
GetHeadPose function can find head pose as long the head is facing the camera to some degree. It fails when on images where at least 50% isn't visible.

The function is also capable of isolating the largest head in a provided image and applying the head pose calculation on it.

The main function in GetHeadPose iterates through all the images in the AFLW2000 dataset